+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+
+## 0.0.7 (2021-11-10)
+
+### Changed
+
+- Rename todo -> ...
+
+
 ## 0.0.6 (2021-11-03)
 
 ### Changed